[Silver III] Title: 주유소, Time: 304 ms, Memory: 41508 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -20,14 +20,14 @@ public class Main {
             cost[i] = Integer.parseInt(st.nextToken());
         }
 
-        int answer = 0;
+        long answer = 0;
         int min = Integer.MAX_VALUE;
         for(int i = 0; i < N - 1; i++) {
             if(min > cost[i]) {
                 min = cost[i];
             }
 
-            answer += min * len[i];
+            answer += (long) min * len[i];
         }
 
         System.out.println(answer);
